feat!: :boom: Add check for formatters and more!
New version!

There's a new FIRST REQUIRED POSITIONAL argument to the cli: task.
It can be format or check.
If it's format, the formatters are expected to MODIFY/FIX the code.
If it's check, the formatters are expected to CHECK the code.
In both cases linters will be ran.
Now the config looks like this for formatters:
{"packages": ["package", ...], format_command: "{pyexe} -m package {path}", check_command: "{pyexe} -m package --check {path}", allow_nonzero_on_format: true}
And for linters: {"packages": ["package", ...], "command": "{pyexe} -m package {path}"}