refactor: :coffin: Remove POTENTIAL_ERRORS
This feature doesn't work. But there is a way to solve it: pass capture_output=True to subprocess.run(), and then check stdout. But if we capture the output, it will not be written to stdout, and we will have to print it later. In my opinion removing this feature is better, than capturing the output.